refactor validator
extract QueryParamsValidator to a different class
