+++ lib/TransactionValidator.js
@@ -2,14 +2,8 @@ const { getBodyMismatchObject, getMismatchWithSuggestedFix } = require('./utils/
   { SOAPMessageHelper } = require('./utils/SOAPMessageHelper'),
   sdk = require('postman-collection'),
   {
-    getAllButProtocolAndHost
-  } = require('./utils/urlUtils'),
-  {
-    URLWithParamsHelper
-  } = require('./utils/URLWithParamsHelper'),
-  {
-    getValueExample
-  } = require('./utils/knownTypes'),
+    QueryParamsValidator
+  } = require('./transactionValidator/QueryParamsValidator'),
   Ajv = require('ajv'),
   {
     getCleanSchema,
@@ -31,7 +25,6 @@ const { getBodyMismatchObject, getMismatchWithSuggestedFix } = require('./utils/
   INVALID_SOAP_METHOD_REASON_CODE = 'INVALID_SOAP_METHOD',
   RESPONSE_BODY_PROPERTY = 'RESPONSE_BODY',
   BODY_PROPERTY = 'BODY',
-  QUERYPARAM_PROPERTY = 'QUERYPARAM',
   RESPONSE_BODY_NOT_PROVIDED_MSG = 'Response body not provided',
   REQUEST_BODY_NOT_PROVIDED_MSG = 'Request body not provided',
   SOAP_METHOD_PROPERTY = 'SOAP_METHOD',
@@ -43,8 +36,7 @@ const { getBodyMismatchObject, getMismatchWithSuggestedFix } = require('./utils/
     MESSAGE_TAG_BODY,
     SOAP_PROTOCOL,
     SOAP12_PROTOCOL,
-    HTTP_PROTOCOL,
-    URL_ENCODED_GET
+    HTTP_PROTOCOL
   } = require('./constants/processConstants'),
   {
     isPmVariable,
@@ -821,77 +813,8 @@ class TransactionValidator {
    * @returns {Array} missmatches array
    */
   validateQueryParams({ itemRequestProcessedInfo, operationFromWSDL, options }) {
-    let rawItemRequestUrl,
-      operationUrlWithoutProtocol,
-      mimeContentInput,
-      transactionPathPrefix = '$.request.url.query',
-      mismatches = [],
-      {
-        showMissingInSchemaErrors,
-        validationPropertiesToIgnore,
-        suggestAvailableFixes
-      } = options;
-
-    if (validationPropertiesToIgnore && validationPropertiesToIgnore.includes(QUERYPARAM_PROPERTY)) {
-      return mismatches;
-    }
-
-    rawItemRequestUrl = itemRequestProcessedInfo.fullURL;
-    operationUrlWithoutProtocol = getAllButProtocolAndHost(operationFromWSDL.url);
-
-    if (rawItemRequestUrl === operationUrlWithoutProtocol) {
-      return mismatches;
-    }
-    mimeContentInput = operationFromWSDL.mimeContentInput;
-    if (mimeContentInput && mimeContentInput.mimeType === URL_ENCODED_GET) {
-      const childrenParams = operationFromWSDL.input[0].children,
-        helper = new URLWithParamsHelper(),
-        keyValuesArray = helper.getKeyValuesFromURLParams(rawItemRequestUrl);
-
-      if (showMissingInSchemaErrors) {
-        keyValuesArray.forEach((keyValue, index) => {
-          let foundOperationParam = childrenParams.find((paramInOperation) => {
-            return keyValue.key === paramInOperation.name;
-          });
-          if (!foundOperationParam) {
-            mismatches.push({
-              property: QUERYPARAM_PROPERTY,
-              transactionJsonPath: transactionPathPrefix + `[${index}]`,
-              schemaJsonPath: null,
-              reasonCode: 'MISSING_IN_SCHEMA',
-              reason: `The query parameter ${keyValue.key} was not found in the schema`
-            });
-          }
-        });
-      }
-      childrenParams.forEach((paramInOperation) => {
-        let foundTransactionParam = keyValuesArray.find((keyValue) => {
-          return keyValue.key === paramInOperation.name;
-        });
-        if (!foundTransactionParam) {
-          let mismatchObj = {
-            property: QUERYPARAM_PROPERTY,
-            transactionJsonPath: transactionPathPrefix,
-            schemaJsonPath: operationFromWSDL.xpathInfo.xpath,
-            reasonCode: 'MISSING_IN_REQUEST',
-            reason: `The required query parameter "${paramInOperation.name}" was not found in the transaction`
-          };
-
-          if (suggestAvailableFixes) {
-            mismatchObj.suggestedFix = {
-              key: paramInOperation.name,
-              actualValue: null,
-              suggestedValue: {
-                key: paramInOperation.name,
-                value: getValueExample(paramInOperation)
-              }
-            };
-          }
-          mismatches.push(mismatchObj);
-        }
-      });
-    }
-    return mismatches;
+    const paramsValidator = new QueryParamsValidator();
+    return paramsValidator.validate({ itemRequestProcessedInfo, operationFromWSDL, options });
   }
 }
