+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "./scripts/test.sh",
-    "unit-tests": "./node_modules/mocha/bin/mocha test/unit/**.test.js --timeout 5000",
+    "unit-tests": "./node_modules/mocha/bin/mocha './test/unit/**/**.test.js'",
     "e2e-tests": "./node_modules/mocha/bin/mocha test/e2e/**.test.js --timeout 5000",
     "convert-validation-tests": "./node_modules/mocha/bin/mocha test/convert-validation/**.test.js --timeout 10000",
     "merge-convert-validate-tests": "./node_modules/mocha/bin/mocha test/merge-convert-validate/**.test.js --timeout 5000",
